Fix compiling with integer cells
In case zfconf.h is modified so that zf_cell is an integer type, the
sscanf() call in zf_host_parse_num() will no longer compile.

This is arguably a host-specific detail, but this commit proposes a
solution wherein the scanf format specifier is also included in the
zfconf.h definitions.