Fixes issue #40
When stubbing and updating jump target addresses in `if`, `else` and
`fi`, variable-length encoding was used; this caused problems when the
stub address and final adress were encoded in different length,
corrupting the dictionary.

This change provides a fix by always using the maximum size
variable-length encoding for the initial stub and final target address.